kind of cracked the support for floats. Kind of.
it's complicated because there are just a few data structures that are
used for everything. I.e. for both the AST and the number of arguments
passed to a function. So there is some casting needed ever now and then.

Still to do: in theory arrays can be indexed by numbers bigger than
ints.

Also, floats are not parsed from the source file yet.